change liquibase to flyway, fix stack overflow with many to many relations

+++ src/main/java/com/example/mybookshopapp/data/BookReviewLikeEntity.java
@@ -12,11 +12,9 @@ public class BookReviewLikeEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
-
+    private Integer id;
     private LocalDateTime time;
-
-    private short value;
+    private Integer value;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "review_id", referencedColumnName = "id")
