+++ src/main/java/com/example/mybookshopapp/data/UserContactEntity.java
@@ -12,18 +12,12 @@ public class UserContactEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
-
+    private Integer id;
     private ContactType type;
-
-    private short approved;
-
+    private Integer approved;
     private String code;
-
-    private int codeTrials;
-
+    private Integer codeTrials;
     private LocalDateTime codeTime;
-
     private String contact;
 
     @ManyToOne
